A Couple of Changes
Hey that's nice work!

Just a couple of notes. `namespace std` has a bunch of functions in it. When you write `using namespace std`, you are importing _all_ of them! Imagine a large project with a million lines of code. The chances of writing a function with the same one in the `std namespace` can be pretty high.  It is better form to instead include only the functions that you need from that namespace. You could be even more particular and leave out the 
`using std::cout;`
line that I put in and instead put `std::` in front of every `cout` and `cin`. 

I also switched from using arrays to using `std::vector` instead. They are re-sizable which means someone can enter any number of items in the list to be sorted. They are also preferred to use in many cases because they are safer than plain arrays.

<h1>Take Values by Reference</h1>
Even though you're passing the array to the functions, it's making a copy of it in the function. That means when the function returns, the array that you passed in is unchanged! Every time you write a function like
`void myFunction(int myInt)`
and call it like
`
int myinteger=10;
myFunction(myInteger);
`
You are creating a copy of myInteger and calling it myInt. 

Because making a copy is creating duplicated stuff in RAM, it is important not to do that unless a copy is needed. Since the vectors might be very large, if we want to pass it to a function but not change it (like in `displayArray`) we can pass it by `const`.