Update main.cpp
When you pass `Card` to displayCard, you were making another copy of `theCard` in memory. Because you are only reading information from `theCard`, there isn't a need to make a second copy in RAM. To use the original `theCard`, we'll take it by reference `&`. Since we aren't changing it, we'll make it `const` too.

This will save RAM and the amount of time to execute the program.